Separate the code in `main`, `renderer` and `util`
