+++ lib/main.js
@@ -1,0 +1,8 @@
+const {checkProcessType} = require('./util');
+
+function init() {
+	checkProcessType('main', 'aperture.main');
+	// TODO
+}
+
+exports.init = init;
